# Stream chunks from blocks-ingester to querier