… CHANGELOG.md
… docs/configuration/v1-guarantees.md
… pkg/chunk/encoding/chunk_test.go
… pkg/ingester/ingester_v2_test.go
… pkg/ingester/lifecycle_test.go
… pkg/querier/batch/batch_test.go
… pkg/querier/batch/chunk_test.go
… pkg/querier/querier_test.go