@pstibrany 2021-03-01
Benchmark comparing old (samples based) and new (chunk based) streaming:

```
name                      old time/op    new time/op    delta
Ingester_v2QueryStream-4    6.21ms ± 1%    1.13ms ± 2%  -81.88%  (p=0.000 n=10+10)

name                      old alloc/op   new alloc/op   delta
Ingester_v2QueryStream-4    3.55MB ± 0%    0.36MB ± 0%  -89.92%  (p=0.000 n=10+10)

name                      old allocs/op  new allocs/op  delta
Ingester_v2QueryStream-4     5.66k ± 0%     3.85k ± 0%  -31.88%  (p=0.000 n=10+10)
```