@pstibrany 2021-03-16
Using this feature makes average time for rule group evaluation to go up:
`avg (cortex_prometheus_rule_group_last_duration_seconds)`